资产系统在 Application 中初始化；添加 AssetManager::CreateAsset 方法，用于创建资产；在 Renderer3D::Init 中创建默认材质文件

+++ Lucky/Source/Lucky/Serialization/SceneSerializer.cpp
@@ -170,7 +170,7 @@ namespace Lucky
                 if (material)
                 {
                     // 通过 Material 实例获取 AssetHandle（Material 继承 Asset，自带 Handle）
-                    out << YAML::Key << "AssetHandle" << YAML::Value << static_cast<uint64_t>(material->GetHandle());
+                    out << YAML::Key << "AssetHandle" << YAML::Value << material->GetHandle();
                 }
                 else
                 {
